fixing tests, commented out some things for now. Test need some things moved around. Currently we are testing Parser methods in OpauthTest
